requirements.txt
Django==4.2
django-widget-tweaks==1.5.0
djangorestframework==3.14.0
django-filter==23.2
